feat: Implement confidence scoring system for parsing
This commit introduces a 'Confidence Scoring System' to assess parsing accuracy.

- Adds `confidence` and `unparsed_text` columns to the `bills` table.
- Refactors `BetCalculator` to calculate a confidence score and return unparsed text.
- Updates backend API to handle the new data.
- Enhances the frontend to display confidence indicators in the bills list.
- Adds a warning and displays unparsed text in the details modal for low-confidence bills.
