+++ frontend/src/components/modals/SettlementModal.jsx
@@ -1,0 +1,220 @@
+import React, { useState, useEffect } from 'react';
+
+// This is a sub-component used only within the SettlementModal.
+function SettlementDetails({ details }) {
+  if (!details) return <div className="details-container">没有详细信息。</div>;
+
+  const { zodiac_bets, number_bets, summary, settlement } = details || {};
+  const safe_zodiac_bets = Array.isArray(zodiac_bets) ? zodiac_bets : [];
+  const safe_number_bets = Array.isArray(number_bets) ? number_bets : [];
+
+  if (safe_zodiac_bets.length === 0 && safe_number_bets.length === 0) {
+    return <div className="details-container">没有解析到投注。</div>;
+  }
+
+  return (
+    <div className="details-container">
+      <table className="settlement-table">
+        <thead>
+          <tr>
+            <th>类型</th>
+            <th>内容</th>
+            <th>金额</th>
+          </tr>
+        </thead>
+        <tbody>
+          {safe_zodiac_bets.map((bet, idx) => (
+            <tr key={`zodiac-${idx}`}>
+              <td className="type-zodiac">生肖投注</td>
+              <td>
+                <span className="zodiac-tag">{bet.zodiac}</span>
+                <span>: {bet.numbers.join(', ')}</span>
+              </td>
+              <td className="amount">{bet.cost} 元</td>
+            </tr>
+          ))}
+          {safe_number_bets.map((bet, idx) => (
+            <tr key={`number-${idx}`}>
+              <td className="type-number">号码投注</td>
+              <td>{bet.numbers.join(', ')}</td>
+              <td className="amount">{bet.cost} 元</td>
+            </tr>
+          ))}
+        </tbody>
+        {summary && (
+          <tfoot>
+            <tr>
+              <td colSpan="2" className="summary-label">号码总数</td>
+              <td className="summary-value">{summary.number_count ?? 0} 个</td>
+            </tr>
+            <tr>
+              <td colSpan="2" className="summary-label">总金额</td>
+              <td className="summary-value">{summary.total_cost ?? 0} 元</td>
+            </tr>
+          </tfoot>
+        )}
+      </table>
+      <div className="settlement-notes-section">
+        <strong>结算备注/说明：</strong>
+        <div className="readonly-notes">
+          {settlement || <span className="no-notes">暂无备注</span>}
+        </div>
+      </div>
+    </div>
+  );
+}
+
+
+function SettlementModal({ open, bill, onClose, onSaveSuccess }) {
+  const [editingSlipIndex, setEditingSlipIndex] = useState(null);
+  const [editedText, setEditedText] = useState('');
+  const [saving, setSaving] = useState(false);
+  const [saveResult, setSaveResult] = useState({ index: null, message: '' });
+
+  useEffect(() => {
+    if (!open) {
+      setEditingSlipIndex(null);
+      setEditedText('');
+      setSaving(false);
+      setSaveResult({ index: null, message: '' });
+    }
+  }, [open]);
+
+  if (!open || !bill) return null;
+
+  let parsedDetails;
+  try {
+    parsedDetails = typeof bill.settlement_details === 'string'
+      ? JSON.parse(bill.settlement_details)
+      : bill.settlement_details;
+  } catch {
+    parsedDetails = { slips: [], summary: {} };
+  }
+
+  const slips = parsedDetails?.slips || [];
+  const summary = parsedDetails?.summary || {};
+
+  const handleEditClick = (index) => {
+    // Note: The structure from the backend is now nested under `settlement_details`
+    setEditingSlipIndex(index);
+    setEditedText(slips[index]?.settlement_details?.settlement || '');
+    setSaveResult({ index: null, message: '' });
+  };
+
+  const handleCancelClick = () => {
+    setEditingSlipIndex(null);
+    setEditedText('');
+  };
+
+  const handleSaveEdit = async () => {
+    setSaving(true);
+    setSaveResult({ index: editingSlipIndex, message: '' });
+    try {
+      const response = await fetch('/update_settlement', {
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify({
+          bill_id: bill.id,
+          slip_index: editingSlipIndex,
+          settlement_text: editedText,
+        }),
+        credentials: 'include'
+      });
+      const data = await response.json();
+      if (data.success) {
+        setSaveResult({ index: editingSlipIndex, message: '保存成功！' });
+        onSaveSuccess();
+        setTimeout(() => setEditingSlipIndex(null), 1500);
+      } else {
+        setSaveResult({ index: editingSlipIndex, message: `保存失败: ${data.error || '未知错误'}` });
+      }
+    } catch (err) {
+      setSaveResult({ index: editingSlipIndex, message: '保存失败: 网络错误' });
+    }
+    setSaving(false);
+  };
+
+  const confidence = bill.confidence ? parseFloat(bill.confidence) : 1.0;
+  const showWarning = confidence < 0.95 && bill.unparsed_text;
+
+  return (
+    <div className="modal-overlay" onClick={onClose}>
+      <div className="modal-content wide-modal" onClick={e => e.stopPropagation()}>
+        <button className="modal-close-button" onClick={onClose}>&times;</button>
+        <h2>结算详情 (账单 #{bill.id})</h2>
+        {showWarning && (
+          <div className="confidence-warning">
+            <strong>⚠️ 系统无法完全识别此邮件，请您仔细核对以下未识别内容：</strong>
+            <pre className="unparsed-text-panel">{bill.unparsed_text}</pre>
+          </div>
+        )}
+        {slips.length === 0 ? (
+          <div className="no-slips-message">没有解析到有效的分段下注单。</div>
+        ) : (
+          <div className="slips-card-container">
+            {slips.map((slip, index) => (
+              <div key={index} className={`bet-slip-card ${editingSlipIndex === index ? 'editing' : ''}`}>
+                <div className="slip-raw">
+                  <div className="slip-card-header">
+                    {slip.region && <span className="region-tag">{slip.region}</span>}
+                    {slip.time ? <span className="time-tag">{slip.time}</span> : `第 ${slip.index} 段`}
+                  </div>
+                  <pre className="slip-pre">{slip.raw}</pre>
+                </div>
+                <div className="slip-result">
+                  <SettlementDetails details={slip.settlement_details} />
+                  {editingSlipIndex === index && (
+                    <div className="editable-notes">
+                      <textarea
+                        value={editedText}
+                        onChange={(e) => setEditedText(e.target.value)}
+                        rows={3}
+                        className="notes-textarea"
+                        placeholder="可编辑结算说明..."
+                        disabled={saving}
+                      />
+                      {saveResult.index === index && saveResult.message && (
+                        <div className={`save-result ${saveResult.message.startsWith('保存成功') ? 'success' : 'error'}`}>
+                          {saveResult.message}
+                        </div>
+                      )}
+                    </div>
+                  )}
+                </div>
+                <div className="slip-cost">
+                  <span>小计</span>
+                  <strong>{slip.settlement_details?.summary?.total_cost || 0} 元</strong>
+                  <div className="slip-actions">
+                    {editingSlipIndex === index ? (
+                      <>
+                        <button onClick={handleSaveEdit} disabled={saving} className="action-button save">
+                          {saving ? '保存中...' : '保存'}
+                        </button>
+                        <button onClick={handleCancelClick} disabled={saving} className="action-button cancel">
+                          取消
+                        </button>
+                      </>
+                    ) : (
+                      <button onClick={() => handleEditClick(index)} className="action-button edit">
+                        编辑备注
+                      </button>
+                    )}
+                  </div>
+                </div>
+              </div>
+            ))}
+            <div className="multi-details-summary">
+              <strong>总计:</strong>
+              <span>{summary.total_cost || 0} 元</span>
+              <span className="summary-divider">|</span>
+              <strong>总号码数:</strong>
+              <span>{summary.total_number_count || 0} 个</span>
+            </div>
+          </div>
+        )}
+      </div>
+    </div>
+  );
+}
+
+export default SettlementModal;
