+++ frontend/src/index.css
@@ -59,3 +59,48 @@ a:hover {
   color: var(--primary-color-hover);
   text-decoration: underline;
 }
+
+/* Status and Confidence Styles */
+.status-processed { color: var(--success-color); font-weight: bold; }
+.status-unrecognized { color: var(--error-color); font-weight: bold; }
+.status-default { color: var(--text-muted-light); }
+
+.confidence-cell {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.confidence-indicator {
+  display: inline-block;
+  width: 12px;
+  height: 12px;
+  border-radius: 50%;
+  flex-shrink: 0;
+}
+.confidence-indicator.high { background-color: #2ecc71; } /* Green */
+.confidence-indicator.medium { background-color: #f1c40f; } /* Yellow */
+.confidence-indicator.low { background-color: #e74c3c; } /* Red */
+
+.confidence-warning {
+  background-color: #fffbe6;
+  border: 1px solid #ffe58f;
+  border-radius: 8px;
+  padding: 12px 16px;
+  margin-bottom: 16px;
+}
+.confidence-warning strong {
+  color: #d46b08;
+}
+
+.unparsed-text-panel {
+  background: #fff;
+  border: 1px dashed #d9d9d9;
+  padding: 8px;
+  margin-top: 8px;
+  white-space: pre-wrap;
+  word-break: break-all;
+  font-size: 0.9em;
+  max-height: 150px;
+  overflow-y: auto;
+}
